backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
psutil==5.9.6
mss==9.0.1
Pillow==10.4.0
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.1.1
psycopg[binary]>=3.2.0
mistralai>=0.1.0
graphviz==0.20.1
